code/www/vpngentor/requirements.txt
anyio
httpcore
httpx
python-dotenv
python-telegram-bot
selenium
urllib3
